Ipk 266 add endpoint for company clients

+++ src/main/java/com/MeetMate/company/CompanyService.java
@@ -1,5 +1,7 @@
 package com.MeetMate.company;
 
+import com.MeetMate.appointment.Appointment;
+import com.MeetMate.appointment.AppointmentRepository;
 import com.MeetMate.company.sequence.CompanySequenceService;
 import com.MeetMate.enums.BusinessType;
 import com.MeetMate.enums.UserRole;
@@ -33,10 +35,50 @@ public class CompanyService {
   private final MongoTemplate mongoTemplate;
   private final CompanySequenceService companySequenceService;
   private final JwtService jwtService;
+  private final AppointmentRepository appointmentRepository;
 
   public Company getCompany(long id) throws IllegalArgumentException {
     return companyRepository.findCompanyById(id)
         .orElseThrow(() -> new EntityNotFoundException("Company not found"));
+  }
+
+  public ArrayList<GetResponse> getClients(String token) throws IllegalAccessException {
+    String email = jwtService.extractUserEmail(token);
+
+    User companyMember = userRepository.findUserByEmail(email)
+        .orElseThrow(() -> new EntityNotFoundException("User not found!"));
+
+    Company company = companyRepository.findCompanyById(companyMember.getAssociatedCompany())
+        .orElseThrow(() -> new EntityNotFoundException("Company not found"));
+
+    if (isNotCompanyOwner(email)
+        && isNotCompanyMember(company, userRepository.findUserByEmail(email)
+        .orElseThrow(() -> new EntityNotFoundException("User not found!"))
+        .getId()))
+      throw new IllegalAccessException("Not a company member");
+
+    ArrayList<Appointment> appointments = appointmentRepository.findAppointmentsByCompanyId(company.getId());
+    ArrayList<User> clients = new ArrayList<>();
+
+    for (Appointment appointment : appointments) {
+      clients.add(
+          userRepository.findUserById(
+              appointment.getClientId()
+          ).orElse(null)
+      );
+    }
+
+    ArrayList<GetResponse> response = new ArrayList<>();
+
+    for (User user : clients) {
+      response.add(GetResponse.builder()
+          .id(user.getId())
+          .name(user.getName())
+          .email(user.getEmail())
+          .build());
+    }
+
+    return response;
   }
 
   public List<Company> getCompanies() {
@@ -67,8 +109,14 @@ public class CompanyService {
   }
 
   @Transactional
-  public void editCompany(String token, String companyName, String description, String businessType) {
-    String ownerEmail = getCompanyWithToken(token).getOwnerEmail();
+  public void editCompany(String token, String companyName, String description, String businessType) throws IllegalAccessException {
+    String ownerEmail = jwtService.extractUserEmail(token);
+
+    if (isNotCompanyOwner(ownerEmail))
+      throw new IllegalAccessException("Not a company owner");
+
+    Company company = companyRepository.findCompanyByOwnerEmail(ownerEmail)
+        .orElseThrow(() -> new EntityNotFoundException("Company not found"));
 
     Query query = new Query(Criteria.where("ownerEmail").is(ownerEmail));
     Update update = new Update();
@@ -79,14 +127,21 @@ public class CompanyService {
   }
 
   @Transactional
-  public void deleteCompany(String token) {
-    Company company = getCompanyWithToken(token);
+  public void deleteCompany(String token) throws IllegalAccessException {
+    String ownerEmail = jwtService.extractUserEmail(token);
+
+    if (isNotCompanyOwner(ownerEmail))
+      throw new IllegalAccessException("Not a company owner");
+
+    Company company = companyRepository.findCompanyByOwnerEmail(ownerEmail)
+        .orElseThrow(() -> new EntityNotFoundException("Company not found"));
+
     try {
       userController.deleteUser("Bearer:" + token);
     } catch (Throwable t) {
       throw new MongoTransactionException("Could not delete company owner");
     }
-    
+
     try {
       for (Long memberId : company.getMemberIds())
         userRepository.deleteById(memberId);
@@ -96,8 +151,19 @@ public class CompanyService {
     companyRepository.delete(company);
   }
 
-  public GetResponse getMember(String token, long memberId) {
-    Company company = getCompanyWithToken(token);
+  ////////////////MEMBER MANAGEMENT////////////////
+
+  public GetResponse getMember(String token, long memberId) throws IllegalAccessException {
+    String email = jwtService.extractUserEmail(token);
+
+    Company company = companyRepository.findCompanyByOwnerEmail(email)
+        .orElseThrow(() -> new EntityNotFoundException("Company not found"));
+
+    if (isNotCompanyOwner(email)
+        && isNotCompanyMember(company, userRepository.findUserByEmail(email)
+        .orElseThrow(() -> new EntityNotFoundException("User not found!"))
+        .getId()))
+      throw new IllegalAccessException("Not a company member");
 
     if (isNotCompanyMember(company, memberId))
       throw new EntityNotFoundException("Not a company member!");
@@ -105,8 +171,17 @@ public class CompanyService {
     return getMemberById(memberId);
   }
 
-  public ArrayList<GetResponse> getAllMembers(String token) {
-    Company company = getCompanyWithToken(token);
+  public ArrayList<GetResponse> getAllMembers(String token) throws IllegalAccessException {
+    String email = jwtService.extractUserEmail(token);
+
+    Company company = companyRepository.findCompanyByOwnerEmail(email)
+        .orElseThrow(() -> new EntityNotFoundException("Company not found"));
+
+    if (isNotCompanyOwner(email)
+        && isNotCompanyMember(company, userRepository.findUserByEmail(email)
+        .orElseThrow(() -> new EntityNotFoundException("User not found!"))
+        .getId()))
+      throw new IllegalAccessException("Not a company member");
 
     ArrayList<GetResponse> members = new ArrayList<>();
     for (Long memberId : company.getMemberIds()) {
@@ -117,8 +192,14 @@ public class CompanyService {
   }
 
   @Transactional
-  public void addMember(String token, String memberEmail, String memberName, String memberPassword) {
-    Company company = getCompanyWithToken(token);
+  public void addMember(String token, String memberEmail, String memberName, String memberPassword) throws IllegalAccessException {
+    String ownerEmail = jwtService.extractUserEmail(token);
+
+    if (isNotCompanyOwner(ownerEmail))
+      throw new IllegalAccessException("Not a company owner");
+
+    Company company = companyRepository.findCompanyByOwnerEmail(ownerEmail)
+        .orElseThrow(() -> new EntityNotFoundException("Company not found"));
 
     MultiValueMap<String, String> memberData = new LinkedMultiValueMap<>();
     memberData.add("email", memberEmail);
@@ -148,7 +229,13 @@ public class CompanyService {
 
   @Transactional
   public void deleteMember(String token, long memberId) throws IllegalAccessException {
-    Company company = getCompanyWithToken(token);
+    String ownerEmail = jwtService.extractUserEmail(token);
+
+    if (isNotCompanyOwner(ownerEmail))
+      throw new IllegalAccessException("Not a company owner");
+
+    Company company = companyRepository.findCompanyByOwnerEmail(ownerEmail)
+        .orElseThrow(() -> new EntityNotFoundException("Company not found"));
 
     if (isNotCompanyMember(company, memberId))
       throw new IllegalAccessException("Not a member of the company");
@@ -167,15 +254,10 @@ public class CompanyService {
     mongoTemplate.updateFirst(query, update, Company.class);
   }
 
-  private Company getCompanyWithToken(String token) throws IllegalArgumentException {
-    String ownerEmail = jwtService.extractUserEmail(token);
-    if (userRepository.findUserByEmail(ownerEmail)
+  private boolean isNotCompanyOwner(String email) {
+    return userRepository.findUserByEmail(email)
         .orElseThrow(() -> new EntityNotFoundException("User not found!"))
-        .getRole() != UserRole.COMPANY_OWNER)
-      throw new IllegalArgumentException("User is not a company owner");
-
-    return companyRepository.findCompanyByOwnerEmail(ownerEmail)
-        .orElseThrow(() -> new EntityNotFoundException("Company not found"));
+        .getRole() != UserRole.COMPANY_OWNER;
   }
 
   private boolean isNotCompanyMember(Company company, long memberId) {
